requirements.txt
asgiref==3.4.1
astroid==2.9.0
attrs==21.4.0
autopep8==1.6.0
coverage==6.1.2
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.0.0
djangorestframework==3.12.4
Faker==11.3.0
gunicorn==20.1.0
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysqlclient==2.1.0
numpy==1.19.5
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pylint==2.12.1
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==2.0.0
typed-ast==1.5.0
typing_extensions==4.0.0
utils==1.0.1
whitenoise==5.3.0
wrapt==1.13.3
